+++ src/main/java/com/jeju/nanaland/domain/review/entity/Review.java
@@ -4,15 +4,15 @@ import com.jeju.nanaland.domain.common.data.Category;
 import com.jeju.nanaland.domain.common.entity.BaseEntity;
 import com.jeju.nanaland.domain.common.entity.Post;
 import com.jeju.nanaland.domain.member.entity.Member;
-import jakarta.persistence.CollectionTable;
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
-import jakarta.persistence.ElementCollection;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.persistence.UniqueConstraint;
 import jakarta.validation.constraints.NotBlank;
@@ -62,8 +62,6 @@ public class Review extends BaseEntity {
   @Column(nullable = false)
   private Integer rating;
 
-  @ElementCollection(targetClass = ReviewKeyword.class)
-  @CollectionTable(name = "review_keywords", joinColumns = @JoinColumn(name = "review_id"))
-  @Enumerated(EnumType.STRING)
+  @OneToMany(mappedBy = "review", cascade = CascadeType.REMOVE)
   private Set<ReviewKeyword> reviewKeywords;
 }
